requirements.txt

# packages:
numpy
pillow


